Remove BGR5 render target test as it is poorly defined
BGR5 includes a "dead" bit and its value is undefined on most APIs for
reads. Applications shouldn't read this bit.

Tests for these kind of formats should be grouped in their on category
on a separate commit. As it is now, we don't want to be testing against
undefined data.